+++ src/test/java/me/karakelley/tictactoe/ApplicationTest.java
@@ -1,9 +1,9 @@
 package me.karakelley.tictactoe;
 
+import me.karakelley.tictactoe.UI.validators.PlayAgainValidator;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 
-import static org.junit.jupiter.api.Assertions.*;
 import static org.mockito.Mockito.*;
 
 class ApplicationTest {
