+++ src/test/java/me/karakelley/tictactoe/PlayAgainValidatorTest.java
@@ -1,8 +1,7 @@
 package me.karakelley.tictactoe;
 
+import me.karakelley.tictactoe.UI.validators.PlayAgainValidator;
 import org.junit.jupiter.api.Test;
-
-import static org.junit.jupiter.api.Assertions.*;
 
 class PlayAgainValidatorTest {
